requirements.txt
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
coverage==5.5
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
distlib==0.3.2
dparse==0.5.1
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Assets==2.0
Flask-Compress==1.10.1
google-api-core==2.0.1
google-api-python-client==2.21.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
pbr==5.6.0
platformdirs==2.3.0
plotly==5.3.1
pluggy==1.0.0
protobuf==3.18.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
safety==1.10.3
six==1.16.0
stevedore==3.4.0
tenacity==8.0.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
virtualenv==20.8.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webassets==2.0
Werkzeug==2.0.1
